Replace model for demo9 start

+++ Blazor/Overview/Demo/Capgemini.Net.Blazor.WebAssembly/Capgemini.Net.Blazor.Components.Demo/AverageRateComponentBase.cs
@@ -10,7 +10,7 @@ namespace Capgemini.Net.Blazor.Components.Demo
         public int MinRate { get; set; }
 
         [Parameter]
-        public int AvgRate { get; set; }
+        public decimal AvgRate { get; set; }
 
         [Parameter]
         public int MaxRate { get; set; }
